Small nits
* Added sidebar menu labels for assistants and chat logs
* Fixed some lint warnings
